Provide a template for building a React SPA with React Router and Material UI.
