+++ app/package.json
@@ -7,7 +7,6 @@
   "dependencies": {
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
-    "firebase": "^8.3.1",
     "formik": "^2.2.6",
     "latest": "^0.2.0",
     "lodash.capitalize": "^4.2.1",
@@ -16,7 +15,6 @@
     "lodash.merge": "^4.6.2",
     "lodash.size": "^4.2.0",
     "react": "^16.14.0",
-    "react-beautiful-dnd": "^13.1.0",
     "react-dom": "^16.14.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.4",
@@ -24,7 +22,6 @@
   },
   "devDependencies": {
     "eslint-config-prettier": "^8.1.0",
-    "firebase-tools": "^9.6.1",
     "prettier": "^2.2.1",
     "prop-types": "^15.7.2"
   },
