+++ app/src/app.jsx
@@ -1,12 +1,12 @@
 import React from "react";
-import { Route, Switch } from "react-router-dom";
+import { Redirect, Route, Switch } from "react-router-dom";
 import { Box } from "@material-ui/core";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import { createMuiTheme, ThemeProvider } from "@material-ui/core/styles";
 
 import CustomTheme from "./assets/theme.json";
 import PageHeader from "./components/page-header";
-import ContactPage from "./pages/contact/contact-us";
+import OrderIntake from "./pages/order/invoke";
 
 const theme = createMuiTheme(CustomTheme);
 
@@ -18,8 +18,9 @@ function App() {
         <PageHeader />
         <Box flexGrow={1} flexShrink={0}>
           <Switch>
-            <Route path="/" exact>
-              <ContactPage />
+            <Redirect exact from="/" to="/order" />
+            <Route path="/order">
+              <OrderIntake />
             </Route>
           </Switch>
         </Box>
